requirements.txt
# Email bot dependencies
# Add more as needed for advanced features
imapclient
pyzmail36
email-validator
